wenn cover in id3 kein jpg fehlermeldung gefixt

+++ js/functions.php
@@ -66,11 +66,11 @@ function getgenrefromalbumID ($id) {
 function coverinmysql ($file, $albumID) {
 
 		require_once './js/thumb/ThumbLib.inc.php';
-		
-				$options = array('jpegQuality' => 70);
-				$optionsbig = array('jpegQuality' => 90);
+				if(mime_content_type($file)=="image/jpeg") {  
+				$options = array('resizeUp' => true, 'jpegQuality' => 60);
+				$optionsbig = array('resizeUp' => true, 'jpegQuality' => 90);
 				
-			//	copy($file, "./tmp/".$albumID."_copy1.jpg");
+				copy($file, "./tmp/".$albumID."_copy1.jpg");
 			//	copy($file, "./tmp/".$albumID."_copy2.jpg");
 			//	copy($file, "./tmp/".$albumID."_copy3.jpg");
 
@@ -80,7 +80,7 @@ function coverinmysql ($file, $albumID) {
 				$thumb = PhpThumbFactory::create($file, $options);
 				$thumb->resize(70, 70)->save('./tmp/'.$albumID.'_kleinesbild.jpg', 'jpg');
 
-				$thumb = PhpThumbFactory::create($file, $optionsbig);
+				$thumb = PhpThumbFactory::create("./tmp/".$albumID."_copy1.jpg", $optionsbig);
 				$thumb->createReflection(10, 40, 90, true, '#a4a4a4')->resize(140, 196)->save('./tmp/'.$albumID.'_reflection.jpg', 'jpg');
 
 				$hndFile = fopen('./tmp/'.$albumID.'_grossesbild.jpg', "r");
@@ -105,15 +105,16 @@ function coverinmysql ($file, $albumID) {
 				unlink('./tmp/'.$albumID.'_kleinesbild.jpg');
 				unlink('./tmp/'.$albumID.'_reflection.jpg');
 			//	unlink('./tmp/'.$albumID.'_copy1.jpg');
-
+					}
 				}
 				
 function thumbnail ($file, $albumID) {
 
 		require_once './js/thumb/ThumbLib.inc.php';
-		
-				$options = array('jpegQuality' => 70);
-				$optionsbig = array('jpegQuality' => 90);
+
+				if(mime_content_type($file)=="image/jpeg") {  
+				$options = array('resizeUp' => true, 'jpegQuality' => 60);
+				$optionsbig = array('resizeUp' => true, 'jpegQuality' => 90);
 				
 			//	copy($file, "./tmp/".$albumID."_copy1.jpg");
 			//	copy($file, "./tmp/".$albumID."_copy2.jpg");
@@ -149,49 +150,50 @@ function thumbnail ($file, $albumID) {
 				unlink('./tmp/'.$albumID.'_kleinesbild.jpg');
 		//		unlink('./tmp/'.$albumID.'_reflection.jpg');
 			//	unlink('./tmp/'.$albumID.'_copy1.jpg');
-
+					}
 				}
 				
 				
 function thumbreflection ($albumID) {
 
 		require_once './js/thumb/ThumbLib.inc.php';
-		
-				$options = array('jpegQuality' => 70);
-				$optionsbig = array('jpegQuality' => 90);
-				
-				$query = "select imgdata,imgtype from album where id=$albumID";
-				$result = @MYSQL_QUERY($query); 
-				$data = @MYSQL_RESULT($result,0,"imgdata"); 
-				$file = './tmp/folder.jpeg';
-				$handle = fopen ($file, 'w+');
-				fwrite($handle, $data);
-				fclose($handle);
+				if(mime_content_type($file)=="image/jpeg") {  
+					$options = array('resizeUp' => true, 'jpegQuality' => 60);
+					$optionsbig = array('resizeUp' => true, 'jpegQuality' => 90);
+					
+					$query = "select imgdata,imgtype from album where id=$albumID";
+					$result = @MYSQL_QUERY($query); 
+					$data = @MYSQL_RESULT($result,0,"imgdata"); 
+					$file = './tmp/folder.jpeg';
+					$handle = fopen ($file, 'w+');
+					fwrite($handle, $data);
+					fclose($handle);
 
-				$thumb = PhpThumbFactory::create($file, $optionsbig);
-				$thumb->createReflection(10, 40, 90, true, '#a4a4a4')->resize(140, 196)->save('./tmp/'.$albumID.'_reflection.jpg', 'jpg');
+					$thumb = PhpThumbFactory::create($file, $optionsbig);
+					$thumb->createReflection(10, 40, 90, true, '#a4a4a4')->resize(140, 196);
+					$thumb->save('./tmp/'.$albumID.'_reflection.jpg', 'jpg');
 
-		//		$hndFile = fopen('./tmp/'.$albumID.'_grossesbild.jpg', "r");
-		//		$data = addslashes(fread($hndFile, filesize('./tmp/'.$albumID.'_grossesbild.jpg')));
+			//		$hndFile = fopen('./tmp/'.$albumID.'_grossesbild.jpg', "r");
+			//		$data = addslashes(fread($hndFile, filesize('./tmp/'.$albumID.'_grossesbild.jpg')));
 
-		//		$hndFilesmall = fopen('./tmp/'.$albumID.'_kleinesbild.jpg', "r");
-		//		$datasmall = addslashes(fread($hndFilesmall, filesize('./tmp/'.$albumID.'_kleinesbild.jpg')));
+			//		$hndFilesmall = fopen('./tmp/'.$albumID.'_kleinesbild.jpg', "r");
+			//		$datasmall = addslashes(fread($hndFilesmall, filesize('./tmp/'.$albumID.'_kleinesbild.jpg')));
 
-				$hndFilebig = fopen('./tmp/'.$albumID.'_reflection.jpg', "r");
-				$databig = addslashes(fread($hndFilebig, filesize('./tmp/'.$albumID.'_reflection.jpg')));
+					$hndFilebig = fopen('./tmp/'.$albumID.'_reflection.jpg', "r");
+					$databig = addslashes(fread($hndFilebig, filesize('./tmp/'.$albumID.'_reflection.jpg')));
 
-				$type = mime_content_type('./tmp/'.$albumID.'_reflection.jpg');
+					$type = mime_content_type('./tmp/'.$albumID.'_reflection.jpg');
 
-		//		mysql_query("UPDATE album SET imgdata = '$data', imgtype = '$type' WHERE id='$albumID'");
-		//		mysql_query("UPDATE album SET imgdata_small = '$datasmall', imgtype = '$type' WHERE id='$albumID'");
-				mysql_query("UPDATE album SET imgdata_big = '$databig', imgtype = '$type' WHERE id='$albumID'");
-				mysql_query("UPDATE album SET coverbig='yes' WHERE id = '$albumID'");
-				
-				unlink($file);
-			//	unlink('./tmp/'.$albumID.'_grossesbild.jpg');
-			//	unlink('./tmp/'.$albumID.'_kleinesbild.jpg');
-				unlink('./tmp/'.$albumID.'_reflection.jpg');
-			//	unlink('./tmp/'.$albumID.'_copy1.jpg');
-
+			//		mysql_query("UPDATE album SET imgdata = '$data', imgtype = '$type' WHERE id='$albumID'");
+			//		mysql_query("UPDATE album SET imgdata_small = '$datasmall', imgtype = '$type' WHERE id='$albumID'");
+					mysql_query("UPDATE album SET imgdata_big = '$databig', imgtype = '$type' WHERE id='$albumID'");
+					mysql_query("UPDATE album SET coverbig='yes' WHERE id = '$albumID'");
+					
+					unlink($file);
+				//	unlink('./tmp/'.$albumID.'_grossesbild.jpg');
+				//	unlink('./tmp/'.$albumID.'_kleinesbild.jpg');
+					unlink('./tmp/'.$albumID.'_reflection.jpg');
+				//	unlink('./tmp/'.$albumID.'_copy1.jpg');
+					}
 				}
 ?>
